Add files via upload
I added the import function to import pandas to be able to run the vlookup.
I was able to name the sample data and call it. I am having issues getting the file to call so I will need to research how to fix this issue. 
I would like to make this code loop and be able to use a vlookup to run different tables at once.